+++ sentry-android-core/src/main/java/io/sentry/android/core/performance/AppStartMetrics.java
@@ -241,6 +241,14 @@ public class AppStartMetrics extends ActivityLifecycleCallbacksAdapter {
     isCallbackRegistered = true;
     appLaunchedInForeground = appLaunchedInForeground || ContextUtils.isForegroundImportance();
     application.registerActivityLifecycleCallbacks(instance);
+    // We post on the main thread a task to post a check on the main thread. On Pixel devices
+    // (possibly others) the first task posted on the main thread is called before the
+    // Activity.onCreate callback. This is a workaround for that, so that the Activity.onCreate
+    // callback is called before the application one.
+    new Handler(Looper.getMainLooper()).post(() -> checkCreateTimeOnMain(application));
+  }
+
+  private void checkCreateTimeOnMain(final @NotNull Application application) {
     new Handler(Looper.getMainLooper())
         .post(
             () -> {