+++ sentry-android-replay/src/main/java/io/sentry/android/replay/ReplayIntegration.kt
@@ -12,7 +12,6 @@ import io.sentry.Integration
 import io.sentry.NoOpReplayBreadcrumbConverter
 import io.sentry.ReplayBreadcrumbConverter
 import io.sentry.ReplayController
-import io.sentry.ScopeObserverAdapter
 import io.sentry.SentryIntegrationPackageStorage
 import io.sentry.SentryLevel.DEBUG
 import io.sentry.SentryLevel.INFO
@@ -21,6 +20,8 @@ import io.sentry.android.replay.capture.BufferCaptureStrategy
 import io.sentry.android.replay.capture.CaptureStrategy
 import io.sentry.android.replay.capture.CaptureStrategy.ReplaySegment
 import io.sentry.android.replay.capture.SessionCaptureStrategy
+import io.sentry.android.replay.gestures.GestureRecorder
+import io.sentry.android.replay.gestures.TouchRecorderCallback
 import io.sentry.android.replay.util.MainLooperHandler
 import io.sentry.android.replay.util.sample
 import io.sentry.android.replay.util.submitSafely
@@ -28,7 +29,6 @@ import io.sentry.cache.PersistingScopeObserver
 import io.sentry.cache.PersistingScopeObserver.BREADCRUMBS_FILENAME
 import io.sentry.cache.PersistingScopeObserver.REPLAY_FILENAME
 import io.sentry.hints.Backfillable
-import io.sentry.protocol.Contexts
 import io.sentry.protocol.SentryId
 import io.sentry.transport.ICurrentDateProvider
 import io.sentry.util.FileUtils
@@ -39,6 +39,7 @@ import java.io.File
 import java.security.SecureRandom
 import java.util.LinkedList
 import java.util.concurrent.atomic.AtomicBoolean
+import kotlin.LazyThreadSafetyMode.NONE
 
 public class ReplayIntegration(
     private val context: Context,
@@ -64,16 +65,20 @@ public class ReplayIntegration(
         recorderConfigProvider: ((configChanged: Boolean) -> ScreenshotRecorderConfig)?,
         replayCacheProvider: ((replayId: SentryId, recorderConfig: ScreenshotRecorderConfig) -> ReplayCache)?,
         replayCaptureStrategyProvider: ((isFullSession: Boolean) -> CaptureStrategy)? = null,
-        mainLooperHandler: MainLooperHandler? = null
+        mainLooperHandler: MainLooperHandler? = null,
+        gestureRecorderProvider: (() -> GestureRecorder)? = null
     ) : this(context, dateProvider, recorderProvider, recorderConfigProvider, replayCacheProvider) {
         this.replayCaptureStrategyProvider = replayCaptureStrategyProvider
         this.mainLooperHandler = mainLooperHandler ?: MainLooperHandler()
+        this.gestureRecorderProvider = gestureRecorderProvider
     }
 
     private lateinit var options: SentryOptions
     private var hub: IHub? = null
     private var recorder: Recorder? = null
+    private var gestureRecorder: GestureRecorder? = null
     private val random by lazy { SecureRandom() }
+    private val rootViewsSpy by lazy(NONE) { RootViewsSpy.install() }
 
     // TODO: probably not everything has to be thread-safe here
     internal val isEnabled = AtomicBoolean(false)
@@ -83,6 +88,7 @@ public class ReplayIntegration(
     private var replayBreadcrumbConverter: ReplayBreadcrumbConverter = NoOpReplayBreadcrumbConverter.getInstance()
     private var replayCaptureStrategyProvider: ((isFullSession: Boolean) -> CaptureStrategy)? = null
     private var mainLooperHandler: MainLooperHandler = MainLooperHandler()
+    private var gestureRecorderProvider: (() -> GestureRecorder)? = null
 
     private lateinit var recorderConfig: ScreenshotRecorderConfig
 
@@ -102,13 +108,8 @@ public class ReplayIntegration(
         }
 
         this.hub = hub
-        this.options.addScopeObserver(object : ScopeObserverAdapter() {
-            override fun setContexts(contexts: Contexts) {
-                // scope screen has fully-qualified name
-                captureStrategy?.onScreenChanged(contexts.app?.viewNames?.lastOrNull()?.substringAfterLast('.'))
-            }
-        })
-        recorder = recorderProvider?.invoke() ?: WindowRecorder(options, this, this, mainLooperHandler)
+        recorder = recorderProvider?.invoke() ?: WindowRecorder(options, this, mainLooperHandler)
+        gestureRecorder = gestureRecorderProvider?.invoke() ?: GestureRecorder(options, this)
         isEnabled.set(true)
 
         try {
@@ -142,7 +143,7 @@ public class ReplayIntegration(
 
         val isFullSession = random.sample(options.experimental.sessionReplay.sessionSampleRate)
         if (!isFullSession && !options.experimental.sessionReplay.isSessionReplayForErrorsEnabled) {
-            options.logger.log(INFO, "Session replay is not started, full session was not sampled and errorSampleRate is not specified")
+            options.logger.log(INFO, "Session replay is not started, full session was not sampled and onErrorSampleRate is not specified")
             return
         }
 
@@ -155,6 +156,7 @@ public class ReplayIntegration(
 
         captureStrategy?.start(recorderConfig)
         recorder?.start(recorderConfig)
+        registerRootViewListeners()
     }
 
     override fun resume() {
@@ -176,8 +178,9 @@ public class ReplayIntegration(
             return
         }
 
-        captureStrategy?.captureReplay(isTerminating == true, onSegmentSent = {
+        captureStrategy?.captureReplay(isTerminating == true, onSegmentSent = { newTimestamp ->
             captureStrategy?.currentSegment = captureStrategy?.currentSegment!! + 1
+            captureStrategy?.segmentTimestamp = newTimestamp
         })
         captureStrategy = captureStrategy?.convert()
     }
@@ -204,7 +207,9 @@ public class ReplayIntegration(
             return
         }
 
+        unregisterRootViewListeners()
         recorder?.stop()
+        gestureRecorder?.stop()
         captureStrategy?.stop()
         isRecording.set(false)
         captureStrategy?.close()
@@ -212,8 +217,10 @@ public class ReplayIntegration(
     }
 
     override fun onScreenshotRecorded(bitmap: Bitmap) {
+        var screen: String? = null
+        hub?.configureScope { screen = it.screen?.substringAfterLast('.') }
         captureStrategy?.onScreenshotRecorded(bitmap) { frameTimeStamp ->
-            addFrame(bitmap, frameTimeStamp)
+            addFrame(bitmap, frameTimeStamp, screen)
         }
     }
 
@@ -255,6 +262,20 @@ public class ReplayIntegration(
 
     override fun onTouchEvent(event: MotionEvent) {
         captureStrategy?.onTouchEvent(event)
+    }
+
+    private fun registerRootViewListeners() {
+        if (recorder is OnRootViewsChangedListener) {
+            rootViewsSpy.listeners += (recorder as OnRootViewsChangedListener)
+        }
+        rootViewsSpy.listeners += gestureRecorder
+    }
+
+    private fun unregisterRootViewListeners() {
+        if (recorder is OnRootViewsChangedListener) {
+            rootViewsSpy.listeners -= (recorder as OnRootViewsChangedListener)
+        }
+        rootViewsSpy.listeners -= gestureRecorder
     }
 
     private fun cleanupReplays(unfinishedReplayId: String = "") {